Update CourseRating class and tests
Update CourseRating class with Data Serialization attributes for better naming - allows for actual property names to be renamed... like 'NumberOfRatings' instead of 'RatersCount' or 'AverageRating' instead of 'Average'. Also means that if the underlying JSON were to change name but the value type was still the same the name of the property would not have to be updated, causing a breaking change.